+++ script/pub/pub.lua
@@ -101,7 +101,7 @@ function m.popReport(brave, name, params)
 end
 
 --- 发布任务
----@parma name string
+---@param name string
 ---@param params any
 ---@return any
 ---@async
@@ -122,7 +122,7 @@ end
 
 --- 发布同步任务，如果任务进入了队列，会返回执行器
 --- 通过 jumpQueue 可以插队
----@parma name string
+---@param name string
 ---@param params any
 ---@param callback function
 function m.task(name, params, callback)